Finished refactor
