--- eeg_infer.py
@@ -1,4 +1,0 @@
-import torch.nn as nn
-import scipy.io as sio
-import numpy as np
-
