--- test_requirements.txt
@@ -1,58 +1,0 @@
-# This file may be used to create an environment using:
-# $ conda create --name <env> --file <this file>
-# platform: osx-64
-blas=1.0=mkl
-brotli=1.0.9=hb1e8313_2
-ca-certificates=2021.10.8=h033912b_0
-certifi=2021.10.8=py39h6e9494a_0
-cycler=0.10.0=pypi_0
-fftw=3.3.8=nompi_h5c49c53_1109
-fonttools=4.25.0=pyhd3eb1b0_0
-freetype=2.10.4=ha233b18_0
-intel-openmp=2021.3.0=hecd8cb5_3375
-jpeg=9d=h9ed2024_0
-kiwisolver=1.3.2=pypi_0
-lcms2=2.12=hf1fd2bf_0
-libcxx=10.0.0=1
-libffi=3.3=hb1e8313_2
-libgfortran=3.0.1=h93005f0_2
-libpng=1.6.37=ha441bb4_0
-libtiff=4.2.0=h87d7836_0
-libuv=1.40.0=haf1e3a3_0
-libwebp-base=1.2.0=h9ed2024_0
-llvm-openmp=12.0.1=hda6cdc1_1
-lz4-c=1.9.3=h23ab428_1
-matplotlib=3.4.3=pypi_0
-matplotlib-base=3.4.2=py39h8b3ea08_0
-mkl=2021.3.0=hecd8cb5_517
-mkl-service=2.4.0=py39h9ed2024_0
-mkl_fft=1.3.0=py39h4a7008c_2
-mkl_random=1.2.2=py39hb2f4e1b_0
-munkres=1.1.4=py_0
-ncurses=6.2=h0a44026_1
-numpy=1.20.3=py39h4b4dc7a_0
-numpy-base=1.20.3=py39he0bd621_0
-olefile=0.46=pyhd3eb1b0_0
-openjpeg=2.4.0=h66ea3da_0
-openssl=1.1.1l=h0d85af4_0
-pillow=8.3.2=pypi_0
-pip=21.2.4=py37hecd8cb5_0
-pyfftw=0.12.0=py39hbb6e609_2
-pyparsing=2.4.7=pyhd3eb1b0_0
-python=3.9.6=h88f2d9e_1
-python-dateutil=2.8.2=pyhd3eb1b0_0
-python_abi=3.9=2_cp39
-pytorch=1.10.0=py3.9_0
-readline=8.1=h9ed2024_0
-scipy=1.6.2=py39hd5f7400_1
-setuptools=52.0.0=py39hecd8cb5_0
-six=1.16.0=pyhd3eb1b0_0
-sqlite=3.36.0=hce871da_0
-tk=8.6.10=hb0a8c7a_0
-tornado=6.1=py39h9ed2024_0
-typing_extensions=3.10.0.2=pyh06a4308_0
-tzdata=2021a=h5d7bf9c_0
-wheel=0.37.0=pyhd3eb1b0_0
-xz=5.2.5=h1de35cc_0
-zlib=1.2.11=h1de35cc_3
-zstd=1.4.9=h322a384_0
